pyproject.toml
[project]
name = "auto-tuner"
version = "0.1.0"
description = "Automatic tunning tool for LLM parameters."
readme = "README.md"
requires-python = ">=3.10,<3.13"
# TODO: fix this. It doesn't load the benchmark correctly if these are not the exact dependencies.
dependencies = [
    "aiofiles==23.2.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "boto3==1.37.21",
    "botocore==1.37.21",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "coloredlogs>=15.0.1",
    "contourpy==1.3.1",
    "cryptography==44.0.2",
    "cycler==0.12.1",
    "deprecated==1.2.18",
    "docker>=7.1.0",
    "events==0.5",
    "fastapi==0.115.12",
    "ffmpy==0.5.0",
    "filelock==3.18.0",
    "fonttools==4.56.0",
    "fsspec==2025.3.0",
    "gradio==5.23.1",
    "gradio-client==1.8.0",
    "groovy==0.1.2",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "huggingface-hub==0.29.3",
    "idna==3.10",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "kiwisolver==1.4.8",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "matplotlib==3.10.1",
    "mdurl==0.1.2",
    "numpy==2.2.4",
    "opensearch-py==2.8.0",
    "orjson==3.10.16",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==10.4.0",
    "pyarrow==19.0.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydub==0.25.1",
    "pygithub==2.6.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pynacl==1.5.0",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rich==13.9.4",
    "ruff==0.11.2",
    "s3transfer==0.11.4",
    "safehttpx==0.1.6",
    "scienceplots==2.1.1",
    "semantic-version==2.10.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.46.1",
    "tomlkit==0.12.0",
    "tqdm==4.67.1",
    "typer==0.15.2",
    "typing-extensions==4.13.0",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "websockets==12.0",
]

[project.scripts]
benchmarker = "benchmarker.launch_bench:main"
dashboard = "dashboard.app:main"
auto-tune = "auto_tune.auto_tune:main"

[tool.ruff]
line-length = 112

[tool.ruff.lint]
select = ["I"]

[tool.ruff.lint.isort]
force-single-line = false
